chore: update Ubuntu version to 22.04 in Docker image
* chore: update the version of Ubuntu to 22

This is mainly to support new standard of C/C++.

* style: add space

* chore: combine RUN commands

This can lead to smaller image size.

* chore: fix missing package

We need libbison-dev for proper `-ly` linking.
The absence of this package leads to a compiler error. It's worth noting
that this package doesn't cause an error in Ubuntu 18, possibly due to
its implicit installation as a dependency of other packages.

* chore: fix missing installation of colorama

Propagated from tmpl-hw3.