Refine specification for better style and cross-referencing
* docs: revise grammar & use format syntax

* docs: update table of contents

This is done with the help of https://github.com/ekalinin/github-markdown-toc.

* docs: add implementation note & format code snippet

* docs: fix typo & use alert block quote

* docs: add reference links

A few sections are moved/added for logically related sections to group
together.

* docs: fix missing semicolon on declarations

* docs: fix incorrect syntax

Adding optional negative sign to the literal constant in the
expressions allow such sign to be applied unlimited times:
```
----1
----expr  # expr is 1
---expr   # expr is -1
--expr    # expr is --1
-expr     # expr is ---1
expr
```
Thus, the optional sign is removed more literal constant. Constant
declaration references to these literals manually.

* docs: add link to the top to the end of each section

* docs: remove the back to top text

To make it not that eye-catching.

* docs: fix broken link

The level-3 sections should be a noun in plural form.

* docs: revise the note

* docs: rearrange examples