docs: defer operator precedence handling to Assignment III
Following student feedback, we've observed that the current implementation
doesn't test the correctness of operator precedence handling. However,
we have encountered a challenge in testing, as Assignment II employs
empty actions for grammar rules, and mismatches in rule matching don't
necessarily disrupt syntax matching. Consequently, there's no observable
behavior to check whether the operations are parsed in the correct precedence.

Hence, we have decided to defer the handling of this matter to
Assignment III. In Assignment III, students will construct the AST, the
structure of which aligns with the precedence of operators, making the
testing process more straightforward.