test: enhance test case for number of parameters exceeding 8
Previously, all arguments passed to the function were set to `2`.
However, in some cases, the test case could pass even if the offset was
miscalculated. For instance, if all arguments were pushed onto the stack
frame in the order of argument 1 through 9, and then the function was
invoked, the 9th argument would be located right at the position pointed
to by the frame pointer. In such a scenario, the 9th argument should be
accessed with offset 0 relative to the frame pointer. It's important to
note that all 9 arguments are stored at offsets 32 through 0 with a
stride of 4 (each argument being 4 bytes in size), and they all have the
value `2`. While the offset could be miscalculated as any of these
offsets and still yield the expected answer, the program's semantics
would be incorrect.

In addition to addressing the offset miscalculation issue, I've removed
irrelevant statements from the test case, simplifying the arguments to
constant values and directly printing out the return value of the
function. Furthermore, I've increased the number of arguments from 9 to
11 to enhance robustness.